Apply PR review feedback: fix GetTheme() init and move RecentProjectEntry to app layer
Agent-Logs-Url: https://github.com/lasrod/assurance-forge/sessions/d2bc9838-2027-4501-872c-7b294b2a1baf
